Revert "Move experimental tests away from TS versions"
This reverts commit 01b9f103249437ad2c3ec605a0c79c69ef0e9743.
