+++ packages/inngest/src/experimental.ts
@@ -1,11 +1,2 @@
 export { getAsyncCtx } from "./components/execution/als.js";
 export type { AsyncContext } from "./components/execution/als.js";
-
-export { channel } from "./experimental/realtime/channel.js";
-export { realtimeMiddleware } from "./experimental/realtime/middleware.js";
-export {
-  getSubscriptionToken,
-  subscribe,
-} from "./experimental/realtime/subscribe.js";
-export { topic } from "./experimental/realtime/topic.js";
-export type { Realtime } from "./experimental/realtime/types.js";
